requirements.txt
flask
flask-cors
flask-bcrypt
mysql-connector-python
